+++ lib/dry/validation/matchers.rb
@@ -3,6 +3,7 @@ require "active_support/core_ext/object/blank"
 require "dry-validation"
 require "dry/validation/matchers/version"
 require "dry/validation/matchers/validate_matcher"
+require "dry/validation/matchers/require_key_matcher"
 require "dry/validation/matchers/integrations/rspec"
 
 module Dry